# docs/swagger.json
{"openapi":"3.0.0","paths":{"/v1/assets":{"get":{"operationId":"AssetsController_getAllAssets","parameters":[{"name":"address","required":true,"in":"query","description":"The ETH address of the owner of the assets.","example":"0x999999cf1046e68e36E1aA2E0E07105eDDD1f08E","schema":{"minLength":42,"maxLength":42,"type":"string"}}],"responses":{"200":{"description":"Returns a list of nfts for Moonbirds collection.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetNFTAssetsResponseDTO"}}}},"500":{"description":"Returns an error if alchemy service breaks."}}}},"/v1/print/validateAndPay":{"post":{"operationId":"PrintController_validateAndPay","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayAndPrintRequestDTO"}}}},"responses":{"200":{"description":"Validate information to pay","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayAndPrintResponseDTO"}}}},"500":{"description":"Returns an error if pay service breaks."}}}},"/v1/print/sendToPrint":{"post":{"operationId":"PrintController_sendToPrint","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendToPrintRequestDTO"}}}},"responses":{"200":{"description":"Print status.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayAndPrintResponseDTO"}}}},"500":{"description":"Returns an error if email service breaks."}}}}},"info":{"title":"Moonbirds | Data API","description":"Moonbirds api for retrieve assets on collection","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"TokenUriDTO":{"type":"object","properties":{"raw":{"type":"string","description":"uri representing the location of the NFTs original metadata blob. This is a backup for you to parse when the metadata field is not automatically populated."},"gateway":{"type":"string","description":"public gateway uri for the raw uri above."}},"required":["raw","gateway"]},"NftMetadataDTO":{"type":"object","properties":{"name":{"type":"string","description":"Name of the NFT asset."},"description":{"type":"string","description":"human-readable description of the NFT asset. (Markdown is supported/rendered on OpenSea and other NFT platforms)."},"image":{"type":"string","description":"URL to the NFT asset image. Can be standard URLs pointing to images on conventional servers, IPFS, or Arweave. Most types of images (SVGs, PNGs, JPEGs, etc.) are supported by NFT marketplaces."},"attributes":{"description":"traits/attributes/characteristics for each NFT asset.","type":"array","items":{"type":"string"}}},"required":["name","description","image","attributes"]},"NftTokenMetadataDTO":{"type":"object","properties":{"tokenType":{"type":"string","enum":["ERC721","ERC1155"],"description":"\"ERC721\" or \"ERC1155\"."}},"required":["tokenType"]},"TokenIDDTO":{"type":"object","properties":{"tokenId":{"type":"string","description":"Id for NFT (hex)."},"tokenMetadata":{"$ref":"#/components/schemas/NftTokenMetadataDTO"}},"required":["tokenId","tokenMetadata"]},"NftContractDTO":{"type":"object","properties":{"address":{"type":"string","description":"address of NFT contract."}},"required":["address"]},"NFTAssetDTO":{"type":"object","properties":{"title":{"type":"string","description":"Name of the NFT asset."},"description":{"type":"string","description":"Brief human-readable description."},"tokenUri":{"$ref":"#/components/schemas/TokenUriDTO"},"media":{"type":"array","items":{"$ref":"#/components/schemas/TokenUriDTO"}},"metadata":{"description":"relevant metadata for NFT contract. This is useful for viewing image url, traits, etc. without having to follow the metadata url in tokenUri to parse manually.","allOf":[{"$ref":"#/components/schemas/NftMetadataDTO"}]},"timeLastUpdated":{"type":"string","description":"Fetch time."},"error":{"type":"string","description":"A string describing a particular reason that we were unable to fetch complete metadata for the NFT."},"balance":{"type":"string","description":"Token balance."},"id":{"description":"Token balance.","allOf":[{"$ref":"#/components/schemas/TokenIDDTO"}]},"contract":{"description":"Token balance.","allOf":[{"$ref":"#/components/schemas/NftContractDTO"}]}},"required":["title","description","tokenUri","media","metadata","timeLastUpdated","error","balance","id","contract"]},"GetNFTAssetsResponseDTO":{"type":"object","properties":{"ownedNfts":{"description":"The list of NFTs.","type":"array","items":{"$ref":"#/components/schemas/NFTAssetDTO"}},"pageKey":{"type":"string","description":"Include pagination key if exists."},"totalCount":{"type":"number","description":"The total number of nfts list (differ from the NFTs length)."}},"required":["ownedNfts","pageKey","totalCount"]},"PayAndPrintRequestDTO":{"type":"object","properties":{"address":{"type":"string","maxLength":42,"minLength":42,"description":"The ETH address of the owner of the assets.","example":"0x999999cf1046e68e36E1aA2E0E07105eDDD1f08E"},"nftAddress":{"type":"string","maxLength":42,"minLength":42,"description":"The ETH address of the NFT.","example":"0x999999cf1046e68e36E1aA2E0E07105eDDD1f08E"},"successURL":{"type":"string","description":"The URL when the payment was correct.","example":"https://example.com/success"},"cancelURL":{"type":"string","description":"The URL when the payment was canceled.","example":"https://example.com/cancel"}},"required":["address","nftAddress","successURL","cancelURL"]},"PayAndPrintResponseDTO":{"type":"object","properties":{"ok":{"type":"boolean","description":"Returns the status of validation, ok if success."},"message":{"type":"string","description":"Returns the validation message for transaction, it will return an error if exists."},"url":{"type":"string","description":"Returns the payment transaction url."},"sessionId":{"type":"string","description":"Returns the session token id"}},"required":["ok","message","url","sessionId"]},"SendToPrintRequestDTO":{"type":"object","properties":{"sessionID":{"type":"string","description":"The payment token id in success page"}},"required":["sessionID"]}}}}